refactor: reorganize pj_marketplace public headers to match project conventions
