+++ pj_marketplace/include/pj_marketplace/marketplace.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "pj_marketplace/download_manager.hpp"
+#include "pj_marketplace/extension.hpp"
+#include "pj_marketplace/extension_detail_dialog.hpp"
+#include "pj_marketplace/extension_manager.hpp"
+#include "pj_marketplace/installed_extension.hpp"
+#include "pj_marketplace/marketplace_window.hpp"
+#include "pj_marketplace/platform_utils.hpp"
+#include "pj_marketplace/registry_manager.hpp"
