+++ pj_marketplace/src/ui/marketplace_window.cpp
@@ -1,9 +1,10 @@
-#include "ui/marketplace_window.hpp"
-#include "ui/extension_detail_dialog.hpp"
+#include "pj_marketplace/marketplace_window.hpp"
+#include "pj_marketplace/extension_detail_dialog.hpp"
 #include "ui_marketplace_window.h"
-#include "core/ExtensionManager.h"
-#include "core/PlatformUtils.h"
-#include "core/RegistryManager.h"
+#include "pj_marketplace/download_manager.hpp"
+#include "pj_marketplace/extension_manager.hpp"
+#include "pj_marketplace/platform_utils.hpp"
+#include "pj_marketplace/registry_manager.hpp"
 
 #include <QComboBox>
 #include <QDialog>
